dashboard/requirements.txt
# Copyright 2023 Google LLC.
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

functions-framework==3.4.*
Flask==2.3.*
google-api-python-client==2.89.*
google-auth==2.20.*
google-auth-httplib2==0.1.*
google-auth-oauthlib==1.0.*
google-cloud-bigquery==3.11.*
pandas==2.0.3
pyarrow
db-dtypes
